refactor: 버그 수정 & guest 페이지 이동 로직 추가
